# [hyperloglog] Improve HLL performance