--- a/hyperloglog.go
+++ b/hyperloglog.go
@@ -66,7 +66,9 @@ func New(registers uint) (*HyperLogLog, error) {
 
 // Reset all internal variables and set the count to zero.
 func (h *HyperLogLog) Reset() {
-	h.Registers = make([]uint8, h.M)
+	for i := range h.Registers {
+		h.Registers[i] = 0
+	}
 }
 
 // Calculate the position of the leftmost 1-bit.