--- a/murmur.go
+++ b/murmur.go
@@ -1,6 +1,8 @@
 package hyperloglog
 
-import "unsafe"
+import (
+	"encoding/binary"
+)
 
 // This file implements the murmur3 32-bit hash on 32bit and 64bit integers
 // for little endian machines only with no heap allocation.  If you are using
@@ -12,13 +14,18 @@ func MurmurString(key string) uint32 {
 	var c1, c2 uint32 = 0xcc9e2d51, 0x1b873593
 	var h, k uint32
 
-	bkey := *(*[]byte)(unsafe.Pointer(&key))
+	bkey := []byte(key)
 	blen := len(bkey)
-	l := blen / 4 // chunk length
 
-	// encode each 4 byte chunk of `key'
+	l := blen / 4 // chunk length
+	tail := bkey[l*4:]
+
+	// for each 4 byte chunk of `key'
 	for i := 0; i < l; i++ {
-		k = *(*uint32)(unsafe.Pointer(&bkey[i*4]))
+		// next 4 byte chunk of `key'
+		k = binary.LittleEndian.Uint32(bkey[i*4:])
+
+		// encode next 4 byte chunk of `key'
 		k *= c1
 		k = (k << 15) | (k >> (32 - 15))
 		k *= c2
@@ -28,23 +35,20 @@ func MurmurString(key string) uint32 {
 	}
 
 	k = 0
-	//remainder
-	if mod := blen % 4; mod > 0 {
-		btail := *(*uint32)(unsafe.Pointer(&bkey[l*4]))
-		switch mod {
-		case 3:
-			k ^= ((btail >> 16) & 0x000000FF) << 16
-			fallthrough
-		case 2:
-			k ^= ((btail >> 8) & 0x000000FF) << 8
-			fallthrough
-		case 1:
-			k ^= btail & 0x000000FF
-			k *= c1
-			k = (k << 15) | (k >> (32 - 15))
-			k *= c2
-			h ^= k
-		}
+	// remainder
+	switch len(tail) {
+	case 3:
+		k ^= uint32(tail[2]) << 16
+		fallthrough
+	case 2:
+		k ^= uint32(tail[1]) << 8
+		fallthrough
+	case 1:
+		k ^= uint32(tail[0])
+		k *= c1
+		k = (k << 15) | (k >> (32 - 15))
+		k *= c2
+		h ^= k
 	}
 
 	h ^= uint32(blen)