What does this PR do?
---------------------
Replaces `rho()` with the stdlib `bits.LeadingZeros32()`, which is intrinsified on amd64 and arm64 and accomplishes the same task (counting the number of leading zeroes in the bit representation of a value) with improved performance. 

Replaces a few calls to the `math` library with bit manipulation, which is less performance intensive

Additional Notes
----------------
Benchmarks indicate significant performance improvements compared to the previous version of HLL: 
```
name                old time/op  new time/op  delta
Reset-10            6.40µs ± 0%  1.36µs ± 0%   ~     (p=1.000 n=1+1)
Count4-10            222ns ± 0%    10ns ± 0%   ~     (p=1.000 n=1+1)
Count5-10            490ns ± 0%    20ns ± 0%   ~     (p=1.000 n=1+1)
Count6-10            915ns ± 0%    46ns ± 0%   ~     (p=1.000 n=1+1)
Count7-10           2.14µs ± 0%  0.10µs ± 0%   ~     (p=1.000 n=1+1)
Count8-10           4.74µs ± 0%  0.21µs ± 0%   ~     (p=1.000 n=1+1)
Count9-10           10.9µs ± 0%   0.5µs ± 0%   ~     (p=1.000 n=1+1)
Count10-10          23.7µs ± 0%   1.0µs ± 0%   ~     (p=1.000 n=1+1)
100MurmurBytes-10   43.0µs ± 0%  42.8µs ± 0%   ~     (p=1.000 n=1+1)
100Murmur64-10       256ns ± 0%   257ns ± 0%   ~     (p=1.000 n=1+1)
100MurmurString-10   629ns ± 0%   629ns ± 0%   ~     (p=1.000 n=1+1)
100Hash32-10         690ns ± 0%   691ns ± 0%   ~     (p=1.000 n=1+1)
MurmurStringBig-10  46.4ms ± 0%  46.7ms ± 0%   ~     (p=1.000 n=1+1)
```
Benchmarks prior to this PR: 
```
BenchmarkReset-10              	  156872	      6405 ns/op
BenchmarkCount4-10             	 5422813	       222.4 ns/op
BenchmarkCount5-10             	 2498756	       490.3 ns/op
BenchmarkCount6-10             	 1299194	       915.0 ns/op
BenchmarkCount7-10             	  546396	      2143 ns/op
BenchmarkCount8-10             	  246973	      4740 ns/op
BenchmarkCount9-10             	  109562	     10890 ns/op
BenchmarkCount10-10            	   50743	     23733 ns/op
Benchmark100MurmurBytes-10     	   28096	     43042 ns/op
Benchmark100Murmur64-10        	 4674187	       255.8 ns/op
Benchmark100MurmurString-10    	 1897222	       629.3 ns/op
Benchmark100Hash32-10          	 1743304	       689.8 ns/op
BenchmarkMurmurStringBig-10    	      25	  46412977 ns/op
```
Benchmarks after this PR's changes:
```
BenchmarkReset-10              	  879480	      1356 ns/op
BenchmarkCount4-10             	100000000	        10.21 ns/op
BenchmarkCount5-10             	57705370	        20.36 ns/op
BenchmarkCount6-10             	26007872	        45.84 ns/op
BenchmarkCount7-10             	12631954	        95.46 ns/op
BenchmarkCount8-10             	 5594330	       214.9 ns/op
BenchmarkCount9-10             	 2608581	       459.3 ns/op
BenchmarkCount10-10            	 1259529	       957.8 ns/op
Benchmark100MurmurBytes-10     	   28106	     42812 ns/op
Benchmark100Murmur64-10        	 4693041	       256.9 ns/op
Benchmark100MurmurString-10    	 1902806	       629.2 ns/op
Benchmark100Hash32-10          	 1741650	       691.0 ns/op
BenchmarkMurmurStringBig-10    	      25	  46652778 ns/op
```